# Conflicts:
#	gs-facade/gs-mall-bbb/src/main/java/com/gs/lshly/facade/bbb/controller/h5/trade/BbbH5TradeController.java

+++ gs-biz-support/gs-commodity/gs-commodity-repository/src/main/java/com/gs/lshly/biz/support/commodity/entity/SkuGoodInfo.java
@@ -150,6 +150,6 @@ public class SkuGoodInfo extends Model {
     private Boolean isInMemberGift;
 
     @ApiModelProperty("in会员积分价格")
-    private Double inMemberPointPrice;
+    private BigDecimal inMemberPointPrice;
 
 }
